+++ src/btree/bt_walk.c
@@ -246,7 +246,7 @@ __tree_walk_internal(WT_SESSION_IMPL *session, WT_REF **refp, uint64_t *walkcntp
     WT_REF *couple, *ref, *ref_orig;
     uint64_t restart_sleep, restart_yield;
     uint32_t slot;
-    u_int current_state;
+    uint8_t current_state;
     bool empty_internal, prev, skip;
 
     btree = S2BT(session);
